requirements.txt
Mastodon.py==1.6.2
PyYAML==6.0
